Fixed bug with keyboard frame
Although the keyboard frame was no longer on screen, and the frame when
transformed to the view had no width, it still had a height. Instead of
manually checking for an empty rect every time keyboardFrame is used, I
chose to adopt this solution, which is to globally return an empty rect if
the keyboard is hidden.